+++ Cargo.toml
@@ -8,7 +8,6 @@ default-members = [
     "node/primitives",
     "node/rpc",
     "node/runtime",
-    "node/test-runtime/*",
     "node/cli",
     "node/pallets/*",
     "geode/builder",
@@ -23,7 +22,6 @@ members = [
     "node/primitives",
     "node/rpc",
     "node/runtime",
-    "node/test-runtime/*",
     "node/cli",
     "node/pallets/*",
     "geode/builder",
