+++ node/rpc/Cargo.toml
@@ -29,9 +29,6 @@ pallet-transaction-payment-rpc = { version = "2.0.0", git = "https://github.com/
 substrate-frame-rpc-system = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
 sp-transaction-pool = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
 sc-transaction-pool = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
-sc-consensus-babe = { version = "0.8.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
-sc-consensus-babe-rpc = { version = "0.8.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
-sp-consensus-babe = { version = "0.8.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
 sc-finality-grandpa = { version = "0.8.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
 sc-finality-grandpa-rpc = { version = "0.8.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
 sc-keystore = { version = "2.0.0", git = "https://github.com/paritytech/substrate.git", rev = "4888ac68c8f451b5843ff17135a34ae0f137dabc" }
