--- node/test-runtime/cli/bin/main.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-fn main() -> sc_cli::Result<()> {
-    automata_test_cli::run()
-}
