--- node/test-runtime/cli/build.rs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use substrate_build_script_utils::{generate_cargo_keys, rerun_if_git_head_changed};
-
-fn main() {
-    generate_cargo_keys();
-
-    rerun_if_git_head_changed();
-}
