--- node/test-runtime/cli/src/executor.rs
@@ -1,13 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use sc_executor::native_executor_instance;
-pub use sc_executor::NativeExecutor;
-
-// Declare an instance of the native executor named `Executor`. Include the wasm binary as the
-// equivalent wasm code.
-native_executor_instance!(
-    pub Executor,
-    automata_test_runtime::api::dispatch,
-    automata_test_runtime::native_version,
-    frame_benchmarking::benchmarking::HostFunctions,
-);
