--- node/test-runtime/cli/src/lib.rs
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod chain_spec;
-mod executor;
-
-#[macro_use]
-mod service;
-#[cfg(feature = "cli")]
-mod cli;
-#[cfg(feature = "cli")]
-mod command;
-
-#[cfg(feature = "cli")]
-pub use cli::*;
-#[cfg(feature = "cli")]
-pub use command::*;
