--- node/test-runtime/cli/src/service.rs
@@ -1,479 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-#![warn(unused_extern_crates)]
-
-use crate::executor::Executor;
-use automata_primitives::Block;
-use automata_test_runtime::RuntimeApi;
-use fc_consensus::FrontierBlockImport;
-use fc_rpc_core::types::PendingTransactions;
-use futures::prelude::*;
-use grandpa::{self};
-use sc_client_api::{ExecutorProvider, RemoteBackend};
-use sc_network::{Event, NetworkService};
-use sc_service::{config::Configuration, error::Error as ServiceError, RpcHandlers, TaskManager};
-use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
-use sp_inherents::InherentDataProviders;
-use sp_runtime::traits::Block as BlockT;
-use std::{
-    collections::HashMap,
-    sync::{Arc, Mutex},
-};
-
-type FullClient = sc_service::TFullClient<Block, RuntimeApi, Executor>;
-type FullBackend = sc_service::TFullBackend<Block>;
-type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
-type FullGrandpaBlockImport =
-    grandpa::GrandpaBlockImport<FullBackend, Block, FullClient, FullSelectChain>;
-type LightClient = sc_service::TLightClient<Block, RuntimeApi, Executor>;
-use sc_telemetry::{TelemetryConnectionNotifier, TelemetrySpan};
-
-pub fn new_partial(
-    config: &Configuration,
-) -> Result<
-    sc_service::PartialComponents<
-        FullClient,
-        FullBackend,
-        FullSelectChain,
-        sp_consensus::DefaultImportQueue<Block, FullClient>,
-        sc_transaction_pool::FullPool<Block, FullClient>,
-        (
-            (
-                sc_consensus_aura::AuraBlockImport<
-                    Block,
-                    FullClient,
-                    FrontierBlockImport<Block, FullGrandpaBlockImport, FullClient>,
-                    AuraPair,
-                >,
-                grandpa::LinkHalf<Block, FullClient, FullSelectChain>,
-            ),
-            Option<TelemetrySpan>,
-        ),
-    >,
-    ServiceError,
-> {
-    let (client, backend, keystore_container, task_manager, telemetry_span) =
-        sc_service::new_full_parts::<Block, RuntimeApi, Executor>(&config)?;
-    let client = Arc::new(client);
-
-    let select_chain = sc_consensus::LongestChain::new(backend.clone());
-
-    let transaction_pool = sc_transaction_pool::BasicPool::new_full(
-        config.transaction_pool.clone(),
-        config.prometheus_registry(),
-        task_manager.spawn_handle(),
-        client.clone(),
-    );
-
-    let (grandpa_block_import, grandpa_link) = grandpa::block_import(
-        client.clone(),
-        &(client.clone() as Arc<_>),
-        select_chain.clone(),
-    )?;
-    let justification_import = grandpa_block_import.clone();
-
-    let frontier_block_import =
-        FrontierBlockImport::new(grandpa_block_import.clone(), client.clone(), true);
-
-    let aura_block_import = sc_consensus_aura::AuraBlockImport::<_, _, _, AuraPair>::new(
-        frontier_block_import.clone(),
-        client.clone(),
-    );
-
-    let inherent_data_providers = sp_inherents::InherentDataProviders::new();
-
-    let import_queue = sc_consensus_aura::import_queue::<_, _, _, AuraPair, _, _>(
-        sc_consensus_aura::slot_duration(&*client)?,
-        aura_block_import.clone(),
-        Some(Box::new(justification_import)),
-        client.clone(),
-        inherent_data_providers.clone(),
-        &task_manager.spawn_handle(),
-        config.prometheus_registry(),
-        sp_consensus::CanAuthorWithNativeVersion::new(client.executor().clone()),
-    )?;
-
-    let import_setup = (aura_block_import, grandpa_link);
-
-    Ok(sc_service::PartialComponents {
-        client,
-        backend,
-        task_manager,
-        keystore_container,
-        select_chain,
-        import_queue,
-        transaction_pool,
-        inherent_data_providers,
-        other: (import_setup, telemetry_span),
-    })
-}
-
-pub struct NewFullBase {
-    pub task_manager: TaskManager,
-    pub inherent_data_providers: InherentDataProviders,
-    pub client: Arc<FullClient>,
-    pub network: Arc<NetworkService<Block, <Block as BlockT>::Hash>>,
-    pub network_status_sinks: sc_service::NetworkStatusSinks<Block>,
-    pub transaction_pool: Arc<sc_transaction_pool::FullPool<Block, FullClient>>,
-}
-
-/// Creates a full service from the configuration.
-pub fn new_full_base(mut config: Configuration) -> Result<NewFullBase, ServiceError> {
-    let sc_service::PartialComponents {
-        client,
-        backend,
-        mut task_manager,
-        import_queue,
-        keystore_container,
-        select_chain,
-        transaction_pool,
-        inherent_data_providers,
-        other: (import_setup, telemetry_span),
-    } = new_partial(&config)?;
-
-    config
-        .network
-        .extra_sets
-        .push(grandpa::grandpa_peers_set_config());
-
-    #[cfg(feature = "cli")]
-    config.network.request_response_protocols.push(
-        sc_finality_grandpa_warp_sync::request_response_config_for_chain(
-            &config,
-            task_manager.spawn_handle(),
-            backend.clone(),
-        ),
-    );
-
-    let (network, network_status_sinks, system_rpc_tx, network_starter) =
-        sc_service::build_network(sc_service::BuildNetworkParams {
-            config: &config,
-            client: client.clone(),
-            transaction_pool: transaction_pool.clone(),
-            spawn_handle: task_manager.spawn_handle(),
-            import_queue,
-            on_demand: None,
-            block_announce_validator_builder: None,
-        })?;
-
-    if config.offchain_worker.enabled {
-        sc_service::build_offchain_workers(
-            &config,
-            backend.clone(),
-            task_manager.spawn_handle(),
-            client.clone(),
-            network.clone(),
-        );
-    }
-
-    let role = config.role.clone();
-    let force_authoring = config.force_authoring;
-    let backoff_authoring_blocks =
-        Some(sc_consensus_slots::BackoffAuthoringOnFinalizedHeadLagging::default());
-    let name = config.network.node_name.clone();
-    let enable_grandpa = !config.disable_grandpa;
-    let prometheus_registry = config.prometheus_registry().cloned();
-    let rpc_client = client.clone();
-    let rpc_network = network.clone();
-    let rpc_transaction_pool = transaction_pool.clone();
-
-    let (rpc_extensions_builder, rpc_setup) = {
-        let (_, grandpa_link) = &import_setup;
-
-        let justification_stream = grandpa_link.justification_stream();
-        let shared_authority_set = grandpa_link.shared_authority_set().clone();
-        let shared_voter_state = grandpa::SharedVoterState::empty();
-        let rpc_setup = shared_voter_state.clone();
-
-        let finality_proof_provider = grandpa::FinalityProofProvider::new_for_service(
-            backend.clone(),
-            rpc_client.clone(),
-            Some(shared_authority_set.clone()),
-        );
-
-        let is_authority = role.is_authority();
-        let subscription_task_executor =
-            sc_rpc::SubscriptionTaskExecutor::new(task_manager.spawn_handle());
-        let pending_transactions: PendingTransactions = Some(Arc::new(Mutex::new(HashMap::new())));
-
-        let rpc_extensions_builder = move |deny_unsafe, subscription_executor| {
-            let pending = pending_transactions.clone();
-            let deps = automata_test_rpc::FullDeps {
-                client: rpc_client.clone(),
-                pool: rpc_transaction_pool.clone(),
-                deny_unsafe,
-                grandpa: automata_test_rpc::GrandpaDeps {
-                    shared_voter_state: shared_voter_state.clone(),
-                    shared_authority_set: shared_authority_set.clone(),
-                    justification_stream: justification_stream.clone(),
-                    subscription_executor,
-                    finality_provider: finality_proof_provider.clone(),
-                },
-                is_authority,
-                enable_dev_signer: false, //TODO
-                network: rpc_network.clone(),
-                pending_transactions: pending.clone(),
-                command_sink: None,
-            };
-
-            automata_test_rpc::create_full(deps, subscription_task_executor.clone())
-        };
-
-        (rpc_extensions_builder, rpc_setup)
-    };
-
-    let (_rpc_handlers, telemetry_connection_notifier) =
-        sc_service::spawn_tasks(sc_service::SpawnTasksParams {
-            config,
-            backend: backend.clone(),
-            client: client.clone(),
-            keystore: keystore_container.sync_keystore(),
-            network: network.clone(),
-            rpc_extensions_builder: Box::new(rpc_extensions_builder),
-            transaction_pool: transaction_pool.clone(),
-            task_manager: &mut task_manager,
-            on_demand: None,
-            remote_blockchain: None,
-            network_status_sinks: network_status_sinks.clone(),
-            system_rpc_tx,
-            telemetry_span,
-        })?;
-
-    let (block_import, grandpa_link) = import_setup;
-
-    if role.is_authority() {
-        let proposer = sc_basic_authorship::ProposerFactory::new(
-            task_manager.spawn_handle(),
-            client.clone(),
-            transaction_pool.clone(),
-            prometheus_registry.as_ref(),
-        );
-
-        let can_author_with =
-            sp_consensus::CanAuthorWithNativeVersion::new(client.executor().clone());
-
-        let aura = sc_consensus_aura::start_aura::<_, _, _, _, _, AuraPair, _, _, _, _>(
-            sc_consensus_aura::slot_duration(&*client)?,
-            client.clone(),
-            select_chain,
-            block_import,
-            proposer,
-            network.clone(),
-            inherent_data_providers.clone(),
-            force_authoring,
-            backoff_authoring_blocks,
-            keystore_container.sync_keystore(),
-            can_author_with,
-        )?;
-
-        // the AURA authoring task is considered essential, i.e. if it
-        // fails we take down the service with it.
-        task_manager
-            .spawn_essential_handle()
-            .spawn_blocking("aura", aura);
-    }
-
-    // Spawn authority discovery module.
-    if role.is_authority() {
-        let authority_discovery_role =
-            sc_authority_discovery::Role::PublishAndDiscover(keystore_container.keystore());
-        let dht_event_stream =
-            network
-                .event_stream("authority-discovery")
-                .filter_map(|e| async move {
-                    match e {
-                        Event::Dht(e) => Some(e),
-                        _ => None,
-                    }
-                });
-        let (authority_discovery_worker, _service) = sc_authority_discovery::new_worker_and_service(
-            client.clone(),
-            network.clone(),
-            Box::pin(dht_event_stream),
-            authority_discovery_role,
-            prometheus_registry.clone(),
-        );
-
-        task_manager.spawn_handle().spawn(
-            "authority-discovery-worker",
-            authority_discovery_worker.run(),
-        );
-    }
-
-    // if the node isn't actively participating in consensus then it doesn't
-    // need a keystore, regardless of which protocol we use below.
-    let keystore = if role.is_authority() {
-        Some(keystore_container.sync_keystore())
-    } else {
-        None
-    };
-
-    let config = grandpa::Config {
-        // FIXME #1578 make this available through chainspec
-        gossip_duration: std::time::Duration::from_millis(333),
-        justification_period: 512,
-        name: Some(name),
-        observer_enabled: false,
-        keystore,
-        is_authority: role.is_network_authority(),
-    };
-
-    if enable_grandpa {
-        // start the full GRANDPA voter
-        // NOTE: non-authorities could run the GRANDPA observer protocol, but at
-        // this point the full voter should provide better guarantees of block
-        // and vote data availability than the observer. The observer has not
-        // been tested extensively yet and having most nodes in a network run it
-        // could lead to finality stalls.
-
-        let shared_voter_state = rpc_setup;
-        let grandpa_config = grandpa::GrandpaParams {
-            config,
-            link: grandpa_link,
-            network: network.clone(),
-            telemetry_on_connect: telemetry_connection_notifier.map(|x| x.on_connect_stream()),
-            voting_rule: grandpa::VotingRulesBuilder::default().build(),
-            prometheus_registry,
-            shared_voter_state,
-        };
-
-        // the GRANDPA voter task is considered infallible, i.e.
-        // if it fails we take down the service with it.
-        task_manager
-            .spawn_essential_handle()
-            .spawn_blocking("grandpa-voter", grandpa::run_grandpa_voter(grandpa_config)?);
-    }
-
-    network_starter.start_network();
-    Ok(NewFullBase {
-        task_manager,
-        inherent_data_providers,
-        client,
-        network,
-        network_status_sinks,
-        transaction_pool,
-    })
-}
-
-/// Builds a new service for a full client.
-pub fn new_full(config: Configuration) -> Result<TaskManager, ServiceError> {
-    new_full_base(config).map(|NewFullBase { task_manager, .. }| task_manager)
-}
-
-pub fn new_light_base(
-    mut config: Configuration,
-) -> Result<
-    (
-        TaskManager,
-        RpcHandlers,
-        Option<TelemetryConnectionNotifier>,
-        Arc<LightClient>,
-        Arc<NetworkService<Block, <Block as BlockT>::Hash>>,
-        Arc<
-            sc_transaction_pool::LightPool<Block, LightClient, sc_network::config::OnDemand<Block>>,
-        >,
-    ),
-    ServiceError,
-> {
-    let (client, backend, keystore_container, mut task_manager, on_demand, telemetry_span) =
-        sc_service::new_light_parts::<Block, RuntimeApi, Executor>(&config)?;
-
-    config
-        .network
-        .extra_sets
-        .push(grandpa::grandpa_peers_set_config());
-
-    let select_chain = sc_consensus::LongestChain::new(backend.clone());
-
-    let transaction_pool = Arc::new(sc_transaction_pool::BasicPool::new_light(
-        config.transaction_pool.clone(),
-        config.prometheus_registry(),
-        task_manager.spawn_handle(),
-        client.clone(),
-        on_demand.clone(),
-    ));
-
-    let (grandpa_block_import, _) = grandpa::block_import(
-        client.clone(),
-        &(client.clone() as Arc<_>),
-        select_chain.clone(),
-    )?;
-
-    let aura_block_import = sc_consensus_aura::AuraBlockImport::<_, _, _, AuraPair>::new(
-        grandpa_block_import.clone(),
-        client.clone(),
-    );
-
-    let import_queue = sc_consensus_aura::import_queue::<_, _, _, AuraPair, _, _>(
-        sc_consensus_aura::slot_duration(&*client)?,
-        aura_block_import,
-        Some(Box::new(grandpa_block_import)),
-        client.clone(),
-        InherentDataProviders::new(),
-        &task_manager.spawn_handle(),
-        config.prometheus_registry(),
-        sp_consensus::NeverCanAuthor,
-    )?;
-
-    let (network, network_status_sinks, system_rpc_tx, network_starter) =
-        sc_service::build_network(sc_service::BuildNetworkParams {
-            config: &config,
-            client: client.clone(),
-            transaction_pool: transaction_pool.clone(),
-            spawn_handle: task_manager.spawn_handle(),
-            import_queue,
-            on_demand: Some(on_demand.clone()),
-            block_announce_validator_builder: None,
-        })?;
-    network_starter.start_network();
-
-    if config.offchain_worker.enabled {
-        sc_service::build_offchain_workers(
-            &config,
-            backend.clone(),
-            task_manager.spawn_handle(),
-            client.clone(),
-            network.clone(),
-        );
-    }
-
-    let light_deps = automata_test_rpc::LightDeps {
-        remote_blockchain: backend.remote_blockchain(),
-        fetcher: on_demand.clone(),
-        client: client.clone(),
-        pool: transaction_pool.clone(),
-    };
-
-    let rpc_extensions = automata_test_rpc::create_light(light_deps);
-
-    let (rpc_handlers, telemetry_connection_notifier) =
-        sc_service::spawn_tasks(sc_service::SpawnTasksParams {
-            on_demand: Some(on_demand),
-            remote_blockchain: Some(backend.remote_blockchain()),
-            rpc_extensions_builder: Box::new(sc_service::NoopRpcExtensionBuilder(rpc_extensions)),
-            client: client.clone(),
-            transaction_pool: transaction_pool.clone(),
-            keystore: keystore_container.sync_keystore(),
-            config,
-            backend,
-            network_status_sinks,
-            system_rpc_tx,
-            network: network.clone(),
-            task_manager: &mut task_manager,
-            telemetry_span,
-        })?;
-
-    Ok((
-        task_manager,
-        rpc_handlers,
-        telemetry_connection_notifier,
-        client,
-        network,
-        transaction_pool,
-    ))
-}
-
-/// Builds a new service for a light client.
-pub fn new_light(config: Configuration) -> Result<TaskManager, ServiceError> {
-    new_light_base(config).map(|(task_manager, _, _, _, _, _)| task_manager)
-}
