--- node/test-runtime/rpc/src/lib.rs
@@ -1,215 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use automata_primitives::{AccountId, Balance, Block, BlockNumber, Hash, Index};
-use fc_rpc_core::types::PendingTransactions;
-use jsonrpc_pubsub::manager::SubscriptionManager;
-use sc_client_api::{
-    backend::{AuxStore, Backend, StateBackend, StorageProvider},
-    client::BlockchainEvents,
-};
-use sc_consensus_manual_seal::rpc::{ManualSeal, ManualSealApi};
-use sc_finality_grandpa::{
-    FinalityProofProvider, GrandpaJustificationStream, SharedAuthoritySet, SharedVoterState,
-};
-use sc_finality_grandpa_rpc::GrandpaRpcHandler;
-use sc_network::NetworkService;
-use sc_rpc::SubscriptionTaskExecutor;
-pub use sc_rpc_api::DenyUnsafe;
-use sp_api::ProvideRuntimeApi;
-use sp_block_builder::BlockBuilder;
-use sp_blockchain::{Error as BlockChainError, HeaderBackend, HeaderMetadata};
-use sp_core::H256;
-use sp_runtime::traits::BlakeTwo256;
-use sp_transaction_pool::TransactionPool;
-use std::{fmt, sync::Arc};
-
-/// Light client extra dependencies.
-pub struct LightDeps<C, F, P> {
-    /// The client instance to use.
-    pub client: Arc<C>,
-    /// Transaction pool instance.
-    pub pool: Arc<P>,
-    /// Remote access to the blockchain (async).
-    pub remote_blockchain: Arc<dyn sc_client_api::light::RemoteBlockchain<Block>>,
-    /// Fetcher instance.
-    pub fetcher: Arc<F>,
-}
-
-/// Extra dependencies for GRANDPA
-pub struct GrandpaDeps<B> {
-    /// Voting round info.
-    pub shared_voter_state: SharedVoterState,
-    /// Authority set info.
-    pub shared_authority_set: SharedAuthoritySet<Hash, BlockNumber>,
-    /// Receives notifications about justification events from Grandpa.
-    pub justification_stream: GrandpaJustificationStream<Block>,
-    /// Executor to drive the subscription manager in the Grandpa RPC handler.
-    pub subscription_executor: SubscriptionTaskExecutor,
-    /// Finality proof provider.
-    pub finality_provider: Arc<FinalityProofProvider<B, Block>>,
-}
-
-/// Full client dependencies.
-pub struct FullDeps<C, P, B> {
-    /// The client instance to use.
-    pub client: Arc<C>,
-    /// Transaction pool instance.
-    pub pool: Arc<P>,
-    /// Whether to deny unsafe calls
-    pub deny_unsafe: DenyUnsafe,
-    /// GRANDPA specific dependencies.
-    pub grandpa: GrandpaDeps<B>,
-    /// The Node authority flag
-    pub is_authority: bool,
-    /// Whether to enable dev signer
-    pub enable_dev_signer: bool,
-    /// Network service
-    pub network: Arc<NetworkService<Block, Hash>>,
-    /// Ethereum pending transactions.
-    pub pending_transactions: PendingTransactions,
-    /// Manual seal command sink
-    pub command_sink:
-        Option<futures::channel::mpsc::Sender<sc_consensus_manual_seal::rpc::EngineCommand<Hash>>>,
-}
-
-/// A IO handler that uses all Full RPC extensions.
-pub type IoHandler = jsonrpc_core::IoHandler<sc_rpc::Metadata>;
-
-/// Instantiate all Full RPC extensions.
-pub fn create_full<C, P, B>(
-    deps: FullDeps<C, P, B>,
-    subscription_task_executor: SubscriptionTaskExecutor,
-) -> jsonrpc_core::IoHandler<sc_rpc_api::Metadata>
-where
-    B: Backend<Block> + 'static,
-    B::State: StateBackend<BlakeTwo256>,
-    C: ProvideRuntimeApi<Block> + StorageProvider<Block, B> + AuxStore,
-    C: BlockchainEvents<Block>,
-    C: HeaderBackend<Block> + HeaderMetadata<Block, Error = BlockChainError>,
-    C: Send + Sync + 'static,
-    C::Api: substrate_frame_rpc_system::AccountNonceApi<Block, AccountId, Index>,
-    C::Api: BlockBuilder<Block>,
-    C::Api: pallet_transaction_payment_rpc::TransactionPaymentRuntimeApi<Block, Balance>,
-    C::Api: fp_rpc::EthereumRuntimeRPCApi<Block>,
-    <C::Api as sp_api::ApiErrorExt>::Error: fmt::Debug,
-    P: TransactionPool<Block = Block> + 'static,
-    H256: From<<P as TransactionPool>::Hash>,
-{
-    use fc_rpc::{
-        EthApi, EthApiServer, EthDevSigner, EthPubSubApi, EthPubSubApiServer, EthSigner,
-        HexEncodedIdProvider, NetApi, NetApiServer, Web3Api, Web3ApiServer,
-    };
-
-    use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApi};
-    use substrate_frame_rpc_system::{FullSystem, SystemApi};
-
-    let mut io = jsonrpc_core::IoHandler::default();
-    let FullDeps {
-        client,
-        pool,
-        deny_unsafe,
-        grandpa,
-        is_authority,
-        network,
-        pending_transactions,
-        command_sink,
-        enable_dev_signer,
-    } = deps;
-
-    let GrandpaDeps {
-        shared_voter_state,
-        shared_authority_set,
-        justification_stream,
-        subscription_executor,
-        finality_provider,
-    } = grandpa;
-
-    io.extend_with(SystemApi::to_delegate(FullSystem::new(
-        client.clone(),
-        pool.clone(),
-        deny_unsafe,
-    )));
-
-    io.extend_with(TransactionPaymentApi::to_delegate(TransactionPayment::new(
-        client.clone(),
-    )));
-
-    io.extend_with(sc_finality_grandpa_rpc::GrandpaApi::to_delegate(
-        GrandpaRpcHandler::new(
-            shared_authority_set.clone(),
-            shared_voter_state,
-            justification_stream,
-            subscription_executor,
-            finality_provider,
-        ),
-    ));
-
-    let mut signers = Vec::new();
-    if enable_dev_signer {
-        signers.push(Box::new(EthDevSigner::new()) as Box<dyn EthSigner>);
-    }
-    io.extend_with(EthApiServer::to_delegate(EthApi::new(
-        client.clone(),
-        pool.clone(),
-        automata_test_runtime::TransactionConverter,
-        network.clone(),
-        pending_transactions.clone(),
-        signers,
-        is_authority,
-    )));
-
-    io.extend_with(NetApiServer::to_delegate(NetApi::new(
-        client.clone(),
-        network.clone(),
-    )));
-
-    io.extend_with(Web3ApiServer::to_delegate(Web3Api::new(client.clone())));
-
-    io.extend_with(EthPubSubApiServer::to_delegate(EthPubSubApi::new(
-        pool.clone(),
-        client.clone(),
-        network.clone(),
-        SubscriptionManager::<HexEncodedIdProvider>::with_id_provider(
-            HexEncodedIdProvider::default(),
-            Arc::new(subscription_task_executor),
-        ),
-    )));
-
-    match command_sink {
-        Some(command_sink) => {
-            io.extend_with(
-                // We provide the rpc handler with the sending end of the channel to allow the rpc
-                // send EngineCommands to the background block authorship task.
-                ManualSealApi::to_delegate(ManualSeal::new(command_sink)),
-            );
-        }
-        _ => {}
-    }
-
-    io
-}
-
-/// Instantiate all Light RPC extensions.
-pub fn create_light<C, P, M, F>(deps: LightDeps<C, F, P>) -> jsonrpc_core::IoHandler<M>
-where
-    C: sp_blockchain::HeaderBackend<Block>,
-    C: Send + Sync + 'static,
-    F: sc_client_api::light::Fetcher<Block> + 'static,
-    P: TransactionPool + 'static,
-    M: jsonrpc_core::Metadata + Default,
-{
-    use substrate_frame_rpc_system::{LightSystem, SystemApi};
-
-    let LightDeps {
-        client,
-        pool,
-        remote_blockchain,
-        fetcher,
-    } = deps;
-    let mut io = jsonrpc_core::IoHandler::default();
-    io.extend_with(SystemApi::<Hash, AccountId, Index>::to_delegate(
-        LightSystem::new(client, remote_blockchain, fetcher, pool),
-    ));
-
-    io
-}
