--- node/test-runtime/runtime/src/impls.rs
@@ -1,56 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::{Authorship, Balances, NegativeImbalance};
-use automata_primitives::Balance;
-use frame_support::traits::{Currency, OnUnbalanced};
-use sp_runtime::traits::Convert;
-
-pub struct Author;
-impl OnUnbalanced<NegativeImbalance> for Author {
-    fn on_nonzero_unbalanced(amount: NegativeImbalance) {
-        Balances::resolve_creating(&Authorship::author(), amount);
-    }
-}
-
-/// Struct that handles the conversion of Balance -> `u64`. This is used for staking's election
-/// calculation.
-pub struct CurrencyToVoteHandler;
-
-impl CurrencyToVoteHandler {
-    fn factor() -> Balance {
-        (Balances::total_issuance() / u64::max_value() as Balance).max(1)
-    }
-}
-
-impl Convert<Balance, u64> for CurrencyToVoteHandler {
-    fn convert(x: Balance) -> u64 {
-        (x / Self::factor()) as u64
-    }
-}
-
-impl Convert<u128, Balance> for CurrencyToVoteHandler {
-    fn convert(x: u128) -> Balance {
-        x * Self::factor()
-    }
-}
-
-/// Struct that handles the conversion of Balance -> `u64`. This is used for token
-pub struct CurrencyToTokenHandler;
-
-impl CurrencyToTokenHandler {
-    fn factor() -> Balance {
-        (Balances::total_issuance() / u64::max_value() as Balance).max(1)
-    }
-}
-
-impl Convert<Balance, u64> for CurrencyToTokenHandler {
-    fn convert(x: Balance) -> u64 {
-        (x / Self::factor()) as u64
-    }
-}
-
-impl Convert<u128, Balance> for CurrencyToTokenHandler {
-    fn convert(x: u128) -> Balance {
-        x as Balance
-    }
-}
